feat: Added experimental support for Octopus Home Pro (3 hours dev time)

+++ custom_components/octopus_energy/coordinators/__init__.py
@@ -46,9 +46,9 @@ class BaseCoordinatorResult:
   last_retrieved: datetime
   next_refresh: datetime
   request_attempts: int
-  refresh_rate_in_minutes: int
+  refresh_rate_in_minutes: float
 
-  def __init__(self, last_retrieved: datetime, request_attempts: int, refresh_rate_in_minutes: int):
+  def __init__(self, last_retrieved: datetime, request_attempts: int, refresh_rate_in_minutes: float):
     self.last_retrieved = last_retrieved
     self.request_attempts = request_attempts
     self.next_refresh = calculate_next_refresh(last_retrieved, request_attempts, refresh_rate_in_minutes)
